Added labels for date and time
Date and time now display labels saying "TIME" and "DATE" to differentiate the two. These labels appear when the watch shakes so a user will know which is which (say it is 0120 and 0120 as both the date and the time). It is handled with poorly coded logic that will probably get fixed one day.